Changed name of class
